pyproject.toml
[project]
name = "MyLittleAgent"
version = "0.1.0"
description = "Workflow orchestration runtime for MyLittleAgent"
readme = "README.md"
license = { text = "Apache-2.0" }
classifiers = [
    "License :: OSI Approved :: Apache Software License",
]
requires-python = ">=3.12,<3.13"
authors = [{ name = "MyLittleAgent Team" }]
dependencies = [
    "pyyaml",
    "openai",
    "tenacity",
    "mcp",
    "fastmcp",
    "faiss-cpu",
    "fastapi==0.124.0",
    "click>=8.1.8,<8.3", # pin until click restores postponed annotations
    "uvicorn",
    "websockets",
    "wsproto",
    "pydantic==2.12.5",
    "requests",
    "pytest",
    "ddgs",
    "beautifulsoup4",
    "matplotlib",
    "networkx",
    "cartopy",
    "pandas>=2.3.3",
    "openpyxl>=3.1.2",
    "numpy>=2.3.5",
    "seaborn>=0.13.2",
    "google-genai>=1.52.0",
    "chardet>=5.2.0",
    "pygame>=2.6.1",
    "filelock>=3.20.1",
    "markdown>=3.10",
    "xhtml2pdf>=0.2.17",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    # Upstream SWIG issue in faiss-cpu on Python 3.12; awaiting SWIG 4.4 fix.
    "ignore:builtin type Swig.*:DeprecationWarning",
    "ignore:builtin type swigvarlink.*:DeprecationWarning",
]


[tool.uv]
package = false
